ai-services/ml-assignment/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP & Database
requests==2.31.0
httpx==0.25.0
pymongo==4.6.0

# ML & NLP (for future steps)
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0
